section 5 with identity Server Complete

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -50,12 +51,13 @@ public class AuctionsController: ControllerBase
 
         return _mapper.Map<AuctionDto>(auction);
     }
-
+    
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
         _context.Auctions.Add(auction);
         var newAuction = _mapper.Map<AuctionDto>(auction);
         await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
@@ -68,7 +70,7 @@ public class AuctionsController: ControllerBase
         
         
     }
-
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdataAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -79,6 +81,10 @@ public class AuctionsController: ControllerBase
             return NotFound();
         }
         //TODO: check seller   ==username
+        if (auction.Seller != User.Identity.Name)
+        {
+            return Forbid();
+        }
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -96,7 +102,7 @@ public class AuctionsController: ControllerBase
 
         return BadRequest("Problem saving changes");
     }
-
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -106,6 +112,10 @@ public class AuctionsController: ControllerBase
             return NotFound();
         }
         // TODO check seller username
+        if (auction.Seller != User.Identity.Name)
+        {
+            return Forbid();
+        }
 
         _context.Auctions.Remove(auction);
